requirements.txt
flask
crew
crewai-tools
dotenv
re
